Minor Refactoring with KtLint Formatting
