+++ app/src/main/java/ru/androidschool/intensiv/data/MockRepository.kt
@@ -29,5 +29,4 @@ object MockRepository {
 
         return tvShowsList
     }
-
 }
