+++ app/src/main/java/ru/androidschool/intensiv/data/MockRepositoryMovieDetails.kt
@@ -28,5 +28,4 @@ object MockRepositoryMovieDetails {
             year = 2020
         )
     }
-
 }
