+++ app/src/main/java/ru/androidschool/intensiv/extensions/ImageViewExtensions.kt
@@ -1,16 +1,14 @@
-package ru.androidschool.intensiv.data
+package ru.androidschool.intensiv.extensions
 
 import android.widget.ImageView
 import com.squareup.picasso.Picasso
 
-object ExtensionMethods {
+object ImageViewExtensions {
 
     fun ImageView.loadImage(string: String) {
 
         Picasso.get()
             .load(string)
             .into(this)
-
     }
-
 }
