+++ app/src/main/java/ru/androidschool/intensiv/ui/feed/MovieItem.kt
@@ -4,8 +4,8 @@ import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
 import com.xwray.groupie.kotlinandroidextensions.Item
 import kotlinx.android.synthetic.main.item_with_text.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.ExtensionMethods.loadImage
 import ru.androidschool.intensiv.data.Movie
+import ru.androidschool.intensiv.extensions.ImageViewExtensions.loadImage
 
 class MovieItem(
     private val content: Movie,
@@ -23,6 +23,5 @@ class MovieItem(
 
         // TODO Получать из модели
         viewHolder.image_preview.loadImage("https://m.media-amazon.com/images/M/MV5BYTk3MDljOWQtNGI2My00OTEzLTlhYjQtOTQ4ODM2MzUwY2IwXkEyXkFqcGdeQXVyNTIzOTk5ODM@._V1_.jpg")
-
     }
 }
