+++ app/src/main/java/ru/androidschool/intensiv/ui/tvshows/TvShowsItem.kt
@@ -1,12 +1,11 @@
 package ru.androidschool.intensiv.ui.tvshows
 
-import com.squareup.picasso.Picasso
 import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
 import com.xwray.groupie.kotlinandroidextensions.Item
 import kotlinx.android.synthetic.main.item_tv_shows.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.ExtensionMethods.loadImage
 import ru.androidschool.intensiv.data.TvShow
+import ru.androidschool.intensiv.extensions.ImageViewExtensions.loadImage
 
 class TvShowsItem(
     private val content: TvShow,
