+++ app/src/main/java/ru/androidschool/intensiv/ui/watchlist/MoviePreviewItem.kt
@@ -1,12 +1,11 @@
 package ru.androidschool.intensiv.ui.watchlist
 
-import com.squareup.picasso.Picasso
 import com.xwray.groupie.kotlinandroidextensions.GroupieViewHolder
 import com.xwray.groupie.kotlinandroidextensions.Item
 import kotlinx.android.synthetic.main.item_with_text.*
 import ru.androidschool.intensiv.R
-import ru.androidschool.intensiv.data.ExtensionMethods.loadImage
 import ru.androidschool.intensiv.data.Movie
+import ru.androidschool.intensiv.extensions.ImageViewExtensions.loadImage
 
 class MoviePreviewItem(
     private val content: Movie,
